pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whip-agent"
version = "0.1.0"
description = "Remote control for AI coding agents via Telegram"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["claude", "codex", "ai", "telegram", "automation"]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.scripts]
whip = "whip.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["whip"]
